% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_lw_relationship.R
\name{est_lwr}
\alias{est_lwr}
\title{Calculate W based on L, a, b}
\usage{
est_lwr(L, param = list(a, b))
}
\arguments{
\item{L}{length}

\item{param}{list of a and b}
}
\value{
Westim
}
\description{
Calculate W based on L, a, b
}
