% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_lwr.R
\name{estim_lwr}
\alias{estim_lwr}
\title{Estimate l-w relationship}
\usage{
estim_lwr(dat, sex = c(1, 2), param = list(a = 1e-05, b = 3))
}
\arguments{
\item{dat}{tidy data with Length, Age, Sex}

\item{sex}{1 for male, 2 for female}

\item{param}{initial param list}
}
\value{
tibble with estimated params
}
\description{
Estimate l-w relationship
}
\examples{
\dontrun{
estim_lwr(
dat = lwdata,
sex = c(1, 2),
param = list(a = 1e-5, b = 3))
}
}
