% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lwdata}
\alias{lwdata}
\title{Length weight data of a fish}
\format{
A data frame with e variables:
\describe{
\item{Length}{Length in centimeter}
\item{Weight}{Weight in gram}
\item{Sex}{Sex 1 for Male, 2 for Female}
}
}
\usage{
lwdata
}
\description{
Length weight data of a fish
}
\keyword{datasets}
